BSD style license.
I know that it seems over the top to add a license to a simple code
base like this one. However, it is polite to do so. It lets any
concerned person know that what restrictions you are putting on
the code.
There are a frighteningly large number of license styles to choose
from. Open Source licenses can be browsed at
    http://opensource.org/licenses